refactoring et possibilité de retirer un fichier uploadé

+++ src/app/boarding/components/config-folder/config-folder.component.ts
@@ -22,34 +22,24 @@ export class ConfigFolderComponent implements OnInit {
 
   currentAgent!: UserInterface;
   idAgent!: number;
-
   typeAgent = 'informel'
-  progress_cni_r?: number;
-  progress_cni_v?: number;
-  progress_geoloc?: number;
-  progress_honneur?: number;
-  progress_connaissance?: number;
-  progress_CGU?: number;
-  progress_residence?: number;
-  progress_statut?: number;
-  file_cni_r?: File;
-  file_cni_v?: File;
-  file_geoloc?: File;
-  file_honneur?: File;
-  file_connaissance?: File;
-  file_CGU?: File;
-  file_residence?: File;
-  file_statut?: File;
   currentFile?: File;
-  progress = 5;
   fileMap: Map<FileType, File | undefined> = new Map();
   progressMap = new Map();
   informelMapKeys: FileType[] = ['cni_r', 'cni_v', 'geoloc', 'honneur', 'connaissance', 'CGU'];
   formelMapKeys: FileType[] = [...this.informelMapKeys, 'statut', 'residence'];
-  informelVariableNames: string[] = ['file_cni_r', 'file_cni_v', 'file_geoloc', 'file_honneur',
-    'file_connaissance', 'file_CGU'];
-  formelVariableNames: string[] = [...this.informelVariableNames, 'file_statut', 'file_residence'];
   fileType!: FileType;
+  showDossiersAgent: boolean = false;
+  paragraphMap: { [key in FileType]: string } = {
+    'cni_r': 'Recto de la CNI',
+    'cni_v': 'Verso de la CNI',
+    'geoloc': 'Géolocalisation du point',
+    'honneur': 'Déclaration de l\'honneur',
+    'connaissance': 'Fiche de Connaissance',
+    'CGU': 'Conditions (CGU)',
+    'residence': 'Contrat de location',
+    'statut': 'Statut de l\'entreprise'
+  };
 
   constructor(private route: ActivatedRoute,
               private userService: UserService,
@@ -62,7 +52,7 @@ export class ConfigFolderComponent implements OnInit {
     try {
       this.idAgent = +this.route.snapshot.params["id"];
     } catch ({message}) {
-      console.log(message);
+      console.error(message);
     }
     if (this.idAgent) {
       this.userService.getUser(this.idAgent).pipe(
@@ -79,30 +69,14 @@ export class ConfigFolderComponent implements OnInit {
         },
       });
     }
-    // this.getTheCurrentAgent();
-  }
-
-  async getTheCurrentAgent() {
-    try {
-      this.idAgent = +this.route.snapshot.params["id"];
-      this.currentAgent = await this.userService.getUserByPromise(this.idAgent);
-      this.userService.saveInLocal('agent', JSON.stringify(this.currentAgent));
-    } catch (error) {
-      console.error(error);
-    }
-    if (!this.currentAgent) {
-      this.currentAgent = JSON.parse(this.userService.getLocalValue('agent'));
-    }
   }
 
   getEvent(event: Event) {
-    this.progress = 5;
     let fileType = this.fileType;
-    this.progressMap.set(fileType, 5);
+    this.progressMap.set(fileType, 10);
     const target = event.target as HTMLInputElement
     if (target.files && target.files.length) {
       this.currentFile = target?.files[0];
-      this.getParagraphMessageByType(this.fileType, this.currentFile);
     }
     if (this.currentFile) {
       console.log(" Le type du fichier " + this.currentFile.type);
@@ -114,7 +88,6 @@ export class ConfigFolderComponent implements OnInit {
         this.notify.snackMessage('Ce type de fichier n\'est pas pris en compte', 4000, 'danger');
         this.currentFile = undefined;
         this.fileMap.delete(fileType);
-        this.resetVariables(fileType);
         return;
       }
       if (!this.checkFileSize(fileType, this.currentFile)) {
@@ -122,31 +95,25 @@ export class ConfigFolderComponent implements OnInit {
         this.notify.snackMessage('La taille du fichier ne doit pas dépasser ' + taille + ' MB', 4000, 'danger');
         this.currentFile = undefined;
         this.fileMap.delete(fileType);
-        this.resetVariables(fileType);
         return;
       }
       let type = Typage[fileType];
       this.fileService.uploadFile(this.currentFile, type).subscribe({
         next: value => {
-          this.notify.snackMessage('Upload avec succès ' + value.toString(), 5000, 'success');
-          this.progress = 100;
-          this.getParagraphMessageByType(this.fileType, this.currentFile, 100);
+          this.notify.snackMessage('Upload avec succès' + value.toString(), 5000, 'success');
           this.progressMap.set(fileType, 100);
         },
         error: err => {
           if (err == 'OK') {
-            this.notify.snackMessage('Upload avec succès ', 5000, 'success');
-            this.progress = 100;
-            this.getParagraphMessageByType(this.fileType, this.currentFile, 100);
+            this.notify.snackMessage('Upload avec succès', 5000, 'success');
             this.progressMap.set(fileType, 100);
           } else {
             console.error(err.toString());
-            this.notify.snackMessage('Error while uploading ' + err.message.toString(),
-              5000, 'danger');
+            this.fileMap.delete(fileType);
+            this.notify.snackMessage('Error while uploading ' + err.message, 5000, 'danger');
           }
         },
       });
-      this.resetVariables('default');
     }
   }
 
@@ -155,137 +122,8 @@ export class ConfigFolderComponent implements OnInit {
     this.fileType = fileType;
   }
 
-  getParagraphMessageByType(fileType: FileType, file?: File, progress?: number) {
-    let paragraph: string;
-    switch (fileType) {
-      case "cni_r":
-        paragraph = 'Recto de la CNI';
-        if (progress) {
-          this.progress_cni_r = progress;
-        }
-        if (file)
-          this.file_cni_r = file;
-        break;
-      case "cni_v":
-        paragraph = 'Verso de la CNI';
-        if (progress) {
-          this.progress_cni_v = progress;
-        }
-        if (file)
-          this.file_cni_v = file;
-        break;
-      case "geoloc":
-        paragraph = 'Géolocalisation du point';
-        if (progress) {
-          this.progress_geoloc = progress;
-        }
-        if (file)
-          this.file_geoloc = file;
-        break;
-      case "honneur":
-        paragraph = 'Déclaration de l\'honneur';
-        if (progress) {
-          this.progress_honneur = progress;
-        }
-        if (file)
-          this.file_honneur = file;
-        break;
-      case "connaissance":
-        paragraph = 'Fiche de Connaissance';
-        if (progress) {
-          this.progress_connaissance = progress;
-        }
-        if (file)
-          this.file_connaissance = file;
-        break;
-      case "CGU":
-        paragraph = 'Conditions (CGU)';
-        if (progress) {
-          this.progress_CGU = progress;
-        }
-        if (file)
-          this.file_CGU = file;
-        break;
-      case "residence":
-        paragraph = 'Contrat de location';
-        if (progress) {
-          this.progress_residence = progress;
-        }
-        if (file)
-          this.file_residence = file;
-        break;
-      case "statut":
-        paragraph = 'Statut de l\'entreprise';
-        if (progress) {
-          this.progress_statut = progress;
-        }
-        if (file)
-          this.file_statut = file;
-        break;
-      default:
-        paragraph = 'Fichier';
-        break;
-    }
-    return paragraph;
-  }
-
-  resetVariables(fileType: FileType | 'all' | 'default') {
-    switch (fileType) {
-      case "cni_r":
-        this.file_cni_r = undefined;
-        this.progress_cni_r = 5;
-        break;
-      case "cni_v":
-        this.file_cni_v = undefined;
-        this.progress_cni_v = 5;
-        break;
-      case "geoloc":
-        this.file_CGU = undefined;
-        this.progress_geoloc = 5;
-        break;
-      case "honneur":
-        this.file_honneur = undefined;
-        this.progress_honneur = 5;
-        break;
-      case "connaissance":
-        this.file_connaissance = undefined;
-        this.progress_connaissance = 5;
-        break;
-      case "CGU":
-        this.file_CGU = undefined;
-        this.progress_CGU = 5;
-        break;
-      case "residence":
-        this.file_residence = undefined;
-        this.progress_residence = 5;
-        break;
-      case "statut":
-        this.file_statut = undefined;
-        this.progress_statut = 5;
-        break;
-      case "all":
-        this.file_cni_r = undefined;
-        this.file_residence = undefined;
-        this.file_CGU = undefined;
-        this.file_geoloc = undefined;
-        this.file_connaissance = undefined;
-        this.file_honneur = undefined;
-        this.file_cni_v = undefined;
-        this.file_statut = undefined;
-        this.progress_cni_r = 5;
-        this.progress_cni_v = 5;
-        this.progress_geoloc = 5;
-        this.progress_honneur = 5;
-        this.progress_connaissance = 5;
-        this.progress_CGU = 5;
-        this.progress_residence = 5;
-        this.progress_statut = 5;
-        break;
-      case 'default':
-        this.currentFile = undefined;
-        this.progress = 5;
-        break;
-    }
+  getParagraphMessageByType(fileType: FileType) {
+    return this.paragraphMap[fileType] || '';
   }
 
   getTooltipContent(fileType: 'gotImage' | 'document'): string {
@@ -298,22 +136,6 @@ export class ConfigFolderComponent implements OnInit {
     const value = 100;
     const keys = this.typeAgent === 'informel' ? this.informelMapKeys : this.formelMapKeys;
     return keys.every(key => this.progressMap.get(key) === value);
-    /*const value = 100;
-    if ( this.typeAgent === 'informel') {
-      for (let key of this.informelMapKeys) {
-        if (this.progressMap.get(key) !== value) {
-          return false
-        }
-      }
-      return true;
-    } else {
-      for (let key of this.formelMapKeys) {
-        if (this.progressMap.get(key) !== value) {
-          return false
-        }
-      }
-      return true;
-    }*/
   }
 
   onSaveDossier() {
@@ -331,70 +153,6 @@ export class ConfigFolderComponent implements OnInit {
       });
       this.fileService.saveAllDossier(dossiers).pipe(
         tap(() => {
-          this.notify.snackMessage('Les fichiers ont bien été uploadé ', 4000, "success");
-          const dialogRef = this.dialog.open(SaveDossierComponent, {
-            data: {
-              agentName: this.currentAgent.name,
-              isOK: true
-            },
-            maxWidth: '25rem',
-          });
-
-        }),
-      ).subscribe();
-
-    } else {
-      this.notify.snackMessage('Vous n\'avez pas chargés tous les fichiers', 4000, "danger");
-    }
-  }
-
-  allowSaveDossierTheOldWay(): boolean {
-    const value = 100;
-    const formelVariables = [this.progress_cni_r, this.progress_cni_v, this.progress_geoloc, this.progress_honneur,
-      this.progress_connaissance, this.progress_CGU, this.progress_residence, this.progress_statut];
-    if (this.typeAgent === 'informel') {
-      return (this.progress_cni_r === 100 && this.progress_cni_v === 100 && this.progress_geoloc === 100
-        && this.progress_honneur === 100 && this.progress_connaissance === 100 && this.progress_CGU === 100)
-    } else {
-      return formelVariables.every(variable => variable === value);
-    }
-  }
-
-  onSaveDossierTheOldWay() {
-    if (this.allowSaveDossierTheOldWay()) {
-      let informelFileVariables = [this.file_cni_r, this.file_cni_v, this.file_geoloc, this.file_honneur,
-        this.file_connaissance, this.file_CGU];
-      let formelFileVariables = [...informelFileVariables, this.file_statut, this.file_residence];
-      let dossiers: DossierInterface[] = [];
-      switch (this.typeAgent) {
-        case 'informel':
-          let i = 0;
-          informelFileVariables.forEach(file => {
-            dossiers.push({
-              name: this.getTypeByNameVariable(this.informelVariableNames[i]) + '_' + file!.name,
-              uploadingFile: file!.name,
-              typeFile: this.getTypeByNameVariable(this.informelVariableNames[i]),
-              acces: this.currentAgent,
-            });
-            i++;
-          })
-          break;
-        case 'formel':
-          let j = 0;
-          formelFileVariables.forEach(file => {
-            dossiers.push({
-              name: this.getTypeByNameVariable(this.formelVariableNames[j]) + '_' + file!.name,
-              uploadingFile: file!.name,
-              typeFile: this.getTypeByNameVariable(this.formelVariableNames[j]),
-              acces: this.currentAgent,
-            });
-            j++;
-          })
-          break;
-      }
-      this.fileService.saveAllDossier(dossiers).pipe(
-        tap(() => {
-          this.notify.snackMessage('Les fichiers ont bien été uploadé ', 4000, "success");
           const dialogRef = this.dialog.open(SaveDossierComponent, {
             data: {
               agentName: this.currentAgent.name,
@@ -416,40 +174,24 @@ export class ConfigFolderComponent implements OnInit {
       this.fileService.checkSize(file!, 'notSelfie')
   }
 
-  getTypeByNameVariable(varName: string) {
-    let varTypeName: FileType = <FileType>(varName.slice(varName.indexOf('_') + 1));
-    console.log("Type recuperer ", varTypeName);
-    return Typage[varTypeName];
-  }
-
   removeFile(typeFile: FileType) {
-    this.fileMap.delete(typeFile);
-    this.progressMap.delete(typeFile);
-  }
-
-  getTypeByVariable(myVar: any) {
-    if (myVar == null) {
-      return Typage.cni_r;
-    }
-
-    let varName = Object.keys({myVar})[0];
-    /*let match = myVar.toString().match(/^(?:function|class)\s*([^\s(]+)/);
-    if (match != null) {
-      varName = match[1];
-    }*/
-    console.log("Nom variable ", varName);
-    let varTypeName: FileType = <FileType>(varName.slice(varName.indexOf('_') + 1));
-    console.log("Type recuperer ", varTypeName);
-    return Typage[varTypeName];
-  }
-
-  checkFileType(filetype: FileType, file?: File): boolean {
-    return filetype.includes('cni') ? this.fileService.checkTypeFile(file!.name, 'CNI') :
-      this.fileService.checkTypeFile(file!.name, 'notCNI');
-  }
-
-
-  showPreviewImg(file: File): string {
-    return URL.createObjectURL(file);
+    const fileName = Typage[typeFile] + '_' + this.fileMap.get(typeFile)!.name;
+    this.fileService.removeFile(fileName).pipe(
+      tap((message) => {
+        this.notify.snackMessage(message.toString(), 1500, 'success');
+      })
+    ).subscribe({
+      next: () => {
+        this.progressMap.set(typeFile, 10);
+        this.fileMap.delete(typeFile);
+      },
+      error: err => {
+        if (err == 'OK') {
+          this.notify.snackMessage(`Ficher ${this.fileMap.get(typeFile)!.name} retiré avec succès`, 5000, 'success');
+          this.progressMap.set(typeFile, 10);
+          this.fileMap.delete(typeFile);
+        }
+      }
+    })
   }
 }
